feat: Add support for different audio bitrates and reciters
The code changes introduce support for different audio bitrates and reciters in the Quran.dart library. This allows users to select the desired audio quality and corresponding reciter for their Quran recitation. The `BitRates` enum is added to represent the available bitrates, and a `getBitRateValue` function is implemented to retrieve the value of a specific bitrate. The `avaiableReciters` map is created to store the available reciters for each bitrate. Additionally, helper functions like `getAllBitRateValues` and `getReciterNamesByBitRate` are added to retrieve all available bitrates and reciter names for a given bitrate, respectively.